Giới hạn quyền của Admin/GV

+++ app/Http/Controllers/API/V1/CourseClassController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\CourseClass;
 use Illuminate\Http\Request;
 use App\Http\Resources\CourseClass as CourseClassResource;
+use Gate;
+use Illuminate\Support\Facades\Auth;
 
 class CourseClassController extends Controller
 {
@@ -16,9 +18,15 @@ class CourseClassController extends Controller
      */
     public function index(Request $request)
     {
+        if(Gate::denies('course_classes.viewAny')) return $this->notAuthorized();
         $params = $request->only(['year', 'semester']);
+
+        // Loc theo quyen
+        if(Auth::user()->roles->isAdmin()) $courseClass = CourseClass::with('course');
+        else $courseClass = Auth::user()->teacherProfile->courseClasses()->with('course');
+
         return response()->success(
-            CourseClassResource::collection(CourseClass::where($params)->with('course')->get())
+            CourseClassResource::collection($courseClass->where($params)->get())
         );
     }
 
@@ -30,6 +38,7 @@ class CourseClassController extends Controller
      */
     public function store(Request $request)
     {
+        if(Gate::denies('course_classes.create')) return $this->notAuthorized();
         if($request->validator->fails())
         {
             return response()->error($request->validator->errors()->all(), 422);
@@ -50,6 +59,8 @@ class CourseClassController extends Controller
     public function show($id)
     {
         $courseClass = CourseClass::find($id);
+        if(Gate::denies('course_classes.view', $courseClass)) return $this->notAuthorized();
+
         //$courseClass->getAllStudentsScores();
         return response()->success(new CourseClassResource($courseClass));
     }
@@ -63,6 +74,7 @@ class CourseClassController extends Controller
      */
     public function update(Request $request, CourseClass $courseClass)
     {
+        if(Gate::denies('course_classes.update', $courseClass)) return $this->notAuthorized();
         $courseClass->update($request->all());
         return response()->success(new CourseClassResource($courseClass));
     }
@@ -75,6 +87,7 @@ class CourseClassController extends Controller
      */
     public function destroy(CourseClass $courseClass)
     {
+        if(Gate::denies('course_classes.delete', $courseClass)) return $this->notAuthorized();
         $courseClass->delete();
         return response()->success("", "Đã xoá thành công.");
     }
